Fix Java compiler warnings in SlideExplorer plugin
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12261 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/SlideExplorer/src/org/micromanager/slideexplorer/Canvas.java
@@ -29,8 +29,8 @@ public class Canvas extends ImageCanvas {
     }
 
     protected void scroll(int sx, int sy) {
-        int ox = xSrcStart + (int) (sx);  //convert to offscreen coordinates
-        int oy = ySrcStart + (int) (sy);
+        int ox = xSrcStart + sx;  //convert to offscreen coordinates
+        int oy = ySrcStart + sy;
         int newx = xSrcStart + (xMouseStart - ox);
         int newy = ySrcStart + (yMouseStart - oy);
         srcRect.x = newx;
